pyproject.toml
[tool.poetry]
name = "personal-email-forwarder"
version = "0.1.0"
description = ""
authors = ["Anthony Sandoval <git@sdval.com>"]

[tool.poetry.dependencies]
python = "^3.8"
boto3 = "^1.16.63"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
